+++ projects/angular/components/ui-grid/src/body/ui-grid-column.directive.ts
@@ -14,8 +14,12 @@ import { identifier } from '@uipath/angular/utilities';
 
 import { Subject } from 'rxjs';
 
-import { UiGridDropdownFilterDirective } from '../filters/ui-grid-dropdown-filter.directive';
-import { UiGridSearchFilterDirective } from '../filters/ui-grid-search-filter.directive';
+import {
+    UiGridDropdownFilterDirective,
+} from '../filters/ui-grid-dropdown-filter.directive';
+import {
+    UiGridSearchFilterDirective,
+} from '../filters/ui-grid-search-filter.directive';
 
 /**
  * @ignore