+++ projects/angular/components/ui-password-indicator/src/ui-password-indicator.module.ts
@@ -5,7 +5,9 @@ import { MatListModule } from '@angular/material/list';
 import { MatProgressBarModule } from '@angular/material/progress-bar';
 import { UiNgLetModule } from '@uipath/angular/directives/ui-ng-let';
 
-import { UiPasswordIndicatorComponent } from './ui-password-indicator.component';
+import {
+    UiPasswordIndicatorComponent,
+} from './ui-password-indicator.component';
 
 @NgModule({
     imports: [