+++ projects/angular/directives/ui-content-loader/src/ui-content-loader.directive.spec.ts
@@ -12,7 +12,9 @@ import { By } from '@angular/platform-browser';
 
 import { BehaviorSubject } from 'rxjs';
 
-import { UiContentSpinnerComponent } from './internal/ui-content-spinner.component';
+import {
+    UiContentSpinnerComponent,
+} from './internal/ui-content-spinner.component';
 import { UiContentLoaderModule } from './ui-content-loader.module';
 
 @Component({