+++ projects/angular/directives/ui-content-loader/src/ui-content-loader.module.ts
@@ -2,7 +2,9 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 
-import { UiContentSpinnerComponent } from './internal/ui-content-spinner.component';
+import {
+    UiContentSpinnerComponent,
+} from './internal/ui-content-spinner.component';
 import { UiContentLoaderDirective } from './ui-content-loader.directive';
 
 @NgModule({
@@ -12,9 +14,6 @@ import { UiContentLoaderDirective } from './ui-content-loader.directive';
     ],
     declarations: [
         UiContentLoaderDirective,
-        UiContentSpinnerComponent,
-    ],
-    entryComponents: [
         UiContentSpinnerComponent,
     ],
     exports: [