+++ projects/angular/directives/ui-spinner-button/src/ui-spinner-button.module.ts
@@ -2,7 +2,9 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 
-import { UiButtonProgressSpinnerComponent } from './internal/ui-button-progress-spinner.component';
+import {
+    UiButtonProgressSpinnerComponent,
+} from './internal/ui-button-progress-spinner.component';
 import { UiSpinnerButtonDirective } from './ui-spinner-button.directive';
 
 @NgModule({
@@ -10,7 +12,6 @@ import { UiSpinnerButtonDirective } from './ui-spinner-button.directive';
         CommonModule,
         MatProgressSpinnerModule,
     ],
-    entryComponents: [UiButtonProgressSpinnerComponent],
     declarations: [
         UiButtonProgressSpinnerComponent,
         UiSpinnerButtonDirective,