+++ projects/angular/directives/ui-virtual-scroll-range-loader/src/ui-virtual-scroll-range-loader.module.ts
@@ -2,7 +2,9 @@ import { ScrollingModule } from '@angular/cdk/scrolling';
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
-import { UiVirtualScrollRangeLoaderDirective } from './ui-virtual-scroll-range-loader.directive';
+import {
+    UiVirtualScrollRangeLoaderDirective,
+} from './ui-virtual-scroll-range-loader.directive';
 
 @NgModule({
     imports: [