+++ projects/angular/directives/ui-virtual-scroll-viewport-resize/src/public_api.ts
@@ -1,2 +1,6 @@
-export { UiVirtualScrollViewportResizeDirective } from './ui-virtual-scroll-viewport-resize.directive';
-export { UiVirtualScrollViewportResizeModule } from './ui-virtual-scroll-viewport-resize.module';
+export {
+    UiVirtualScrollViewportResizeDirective,
+} from './ui-virtual-scroll-viewport-resize.directive';
+export {
+    UiVirtualScrollViewportResizeModule,
+} from './ui-virtual-scroll-viewport-resize.module';