+++ projects/angular/directives/ui-virtual-scroll-viewport-resize/src/ui-virtual-scroll-viewport-resize.module.ts
@@ -1,6 +1,8 @@
 import { NgModule } from '@angular/core';
 
-import { UiVirtualScrollViewportResizeDirective } from './ui-virtual-scroll-viewport-resize.directive';
+import {
+    UiVirtualScrollViewportResizeDirective,
+} from './ui-virtual-scroll-viewport-resize.directive';
 
 @NgModule({
     declarations: [UiVirtualScrollViewportResizeDirective],